+++ src/main/java/com/ssafy/freezetag/domain/room/service/response/RoomConnectResponseDto.java
@@ -1,7 +1,6 @@
 package com.ssafy.freezetag.domain.room.service.response;
 
 import com.ssafy.freezetag.domain.message.entity.MessageRedis;
-import com.ssafy.freezetag.domain.room.service.request.OpenviduResponseDto;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -14,7 +13,7 @@ public class RoomConnectResponseDto {
     private final String roomCode;
     private final String roomName;
     private final Integer roomPersonCount;
-    private final List<RoomMemberInfoDto> members;
+    private final List<RoomMemberInfoResponseDto> members;
     private final Long hostId;
     private final OpenviduResponseDto webrtc;
     private final List<MessageRedis> messages;
